api/requirements.txt
fastapi
uvicorn[standard]
azure-storage-blob
pandas
python-dateutil
orjson
ndjson